Run an sbt server and communicate from vim
Instead of using multiprocessing to run sbt from within vim, use
multiprocessing.communicate to run sbt in a server process. Then send
commands to run over that pipe, and have the server call a vim remote
expr after it's done. This does less forking in vim and seems to trip
less mines as a result